[Refactor] 폴더명 maps에서 house로 변경
