+++ src/main/java/com/efub/livin/house/domain/HouseType.java
@@ -1,4 +1,4 @@
-package com.efub.livin.maps.domain;
+package com.efub.livin.house.domain;
 
 public enum HouseType {
     PRIVATE, // 자취방
