+++ src/main/java/com/efub/livin/house/repository/HouseSaveRepository.java
@@ -1,6 +1,6 @@
-package com.efub.livin.maps.repository;
+package com.efub.livin.house.repository;
 
-import com.efub.livin.maps.domain.House;
+import com.efub.livin.house.domain.House;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface HouseSaveRepository extends JpaRepository<House, Long> {
